pc_server/requirements.txt
# Core
flask==3.0.0
flask-socketio==5.3.5
python-socketio[client]==5.10.0
python-dotenv==1.0.0
PyYAML==6.0.1

# AI & NLP
google-generativeai==0.3.2
edge-tts==6.1.9
speechrecognition==3.10.1
PyAudio==0.2.14

# Vision & Processing
opencv-python==4.9.0.80
numpy==1.24.3

# Optional: Advanced SLAM
# slam-toolbox==0.1.0
# open3d==0.18.0
